# backend/src/main/java/com/example/backend/Modules/Auth/DTO/ProfileDTO.java
package com.example.backend.Modules.Auth.DTO;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ProfileDTO {
    private String name;
    private String email;
    private String phone;
    private String profile;
}